# api-gateway/docs/swagger.yaml
basePath: /
definitions:
...
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.Account:
    properties:
      id:
        type: string
      mask:
        type: string
      name:
        type: string
      subtype:
        type: string
      type:
        type: string
      verification_status:
        type: string
    type: object
...
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreatePlaidAccessTokenResponse:
    properties:
      accessToken:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreatePlaidIDVTokenResponse:
    properties:
      idvToken:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreatePlaidLinkTokenResponse:
    properties:
      linkToken:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreateUserAccountResponse:
    properties:
      error:
        type: string
      galileoAccountDetails:
        $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.GalileoAccountDetails'
      galileoAccountStatus:
        type: string
      message:
        type: string
      status:
        type: integer
    type: object
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreateUserRecordResponse:
    properties:
      message:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.GalileoAccountDetails:
    properties:
      cardId:
        type: string
      cardNumber:
        type: string
      cardSecurityCode:
        type: string
    type: object
...
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.GetAccountStatusResponse:
    properties:
      GalileoAccountCreated:
        type: boolean
      isVerified:
        type: boolean
    type: object
...
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.Institution:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.PlaidMetadata:
    properties:
      accounts:
        items:
          $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.Account'
        type: array
      institution:
        $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.Institution'
      link_session_id:
        type: string
    type: object
  github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.SetIdentityVerifiedResponse:
    properties:
      isVerified:
        type: boolean
    type: object
...
  pkg_user-compliance_routes.Address:
    properties:
      address1:
        type: string
      address2:
        type: string
      city:
        type: string
      countryCode:
        type: string
      postalCode:
        type: string
      state:
        type: string
    type: object
  pkg_user-compliance_routes.CreateAccessTokenRequestBody:
    properties:
      metadata:
        $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.PlaidMetadata'
      publicToken:
        type: string
    type: object
  pkg_user-compliance_routes.CreateUserAccountRequestBody:
    properties:
      address:
        $ref: '#/definitions/pkg_user-compliance_routes.Address'
      personalInfo:
        $ref: '#/definitions/pkg_user-compliance_routes.PersonalInfo'
      username:
        type: string
    type: object
  pkg_user-compliance_routes.PersonalInfo:
    properties:
      dateOfBirth:
        type: string
      email:
        type: string
      firstName:
        type: string
      id:
        type: string
      idType:
        type: integer
      lastName:
        type: string
      phoneNumber:
        type: string
    type: object
  pkg_user-compliance_routes.SetIdentityStatusRequestBody:
    properties:
      link_session_id:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is the client facing API for Sharefunds/Quilt/PoolParty
  title: Sharefunds Client Facing API
  version: 1.0.1
paths:
...
  /user-compliance/plaid/create-accesstoken/{UID}:
    post:
      consumes:
      - application/json
      description: Exchanges the public token for an access token.
      operationId: create_access_token
      parameters:
      - description: User ID
        in: path
        name: UID
        required: true
        type: string
      - description: Public Token to exchange for access token
        in: body
        name: PublicToken
        required: true
        schema:
          $ref: '#/definitions/pkg_user-compliance_routes.CreateAccessTokenRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreatePlaidAccessTokenResponse'
      security:
      - ApiKeyAuth: []
      summary: Creates an access token from the Plaid API.
      tags:
      - User & Compliance Service - Plaid
  /user-compliance/plaid/create-idvtoken/{UID}:
    post:
      consumes:
      - application/json
      description: Creates a link token for the Identity Verification flow, which
        allows the user to verify their identity.
      operationId: create_IDV_token
      parameters:
      - description: User ID
        in: path
        name: UID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreatePlaidIDVTokenResponse'
      security:
      - ApiKeyAuth: []
      summary: Creates an Identity Verification Link Token from the Plaid API.
      tags:
      - User & Compliance Service - Plaid
  /user-compliance/plaid/create-linktoken/{UID}:
    post:
      consumes:
      - application/json
      description: Creates a link token from the Plaid API.
      operationId: create_link_token
      parameters:
      - description: User ID
        in: path
        name: UID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreatePlaidLinkTokenResponse'
      security:
      - ApiKeyAuth: []
      summary: Creates a link token from the Plaid API.
      tags:
      - User & Compliance Service - Plaid
  /user-compliance/users/create-account/{UID}:
    post:
      consumes:
      - application/json
      description: Creates a user's Galileo account, and account record within our
        user-compliance service DB with provided info, returns the Galileo account
        info (Create a new card program) and status.
      operationId: create_user_account
      parameters:
      - description: User ID
        in: path
        name: UID
        required: true
        type: string
      - description: Request Body to create user account
        in: body
        name: AccountInfo
        required: true
        schema:
          $ref: '#/definitions/pkg_user-compliance_routes.CreateUserAccountRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreateUserAccountResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a user's digital wallet (galileo) account, and a record of the
        users info on the user-compliance service DB.
      tags:
      - User & Compliance Service
  /user-compliance/users/create-user-record/{id}:
    post:
      consumes:
      - application/json
      description: Creates a user record on the user-compliance service DB.
      operationId: create_user_record
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.CreateUserRecordResponse'
      security:
      - ApiKeyAuth: []
      summary: Creates a user record on the user-compliance service DB.
      tags:
      - User & Compliance Service
  /user-compliance/users/get-account-status/{UID}:
    get:
      consumes:
      - application/json
      description: Gets a boolean value indicating whether the user has completed
        the identity verification flow, and galileo account creation.
      operationId: get_account_status
      parameters:
      - description: User ID
        in: path
        name: UID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.GetAccountStatusResponse'
      security:
      - ApiKeyAuth: []
      summary: Gets a user's account status, which is stored on the user-compliance
        service DB.
      tags:
      - User & Compliance Service
...
  /user-compliance/users/set-verification-status/{UID}:
    post:
      consumes:
      - application/json
      description: Sets a boolean value indicating whether the user has completed
        the identity verification flow.
      operationId: set_identity_status
      parameters:
      - description: User ID
        in: path
        name: UID
        required: true
        type: string
      - description: Session ID of the link token
        in: body
        name: LinkSessionId
        required: true
        schema:
          $ref: '#/definitions/pkg_user-compliance_routes.SetIdentityStatusRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_Sharefunds_api-gateway_pkg_user-compliance_pb.SetIdentityVerifiedResponse'
      security:
      - ApiKeyAuth: []
      summary: Sets a user's identity status, which is stored on the user-compliance
        service DB.
      tags:
      - User & Compliance Service
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"